Kub : url enforcing
