+++ onyxia-api/src/main/resources/application.properties
@@ -25,6 +25,7 @@ mesos.url=
 # Kubernetes
 kubernetes.enabled=false
 kubernetes.namespace.prefix=user-
+kubernetes.publish.domain=selfservice-kub.example.com
 
 # Catalogs
 catalogs.configuration=classpath:multiverse.json
